Move all code into lib folder and replace relative_require
