+++ lib/certificator.rb
@@ -1,8 +1,8 @@
 # encoding:UTF-8
 require 'certificator/version'
-require 'configuration/certificate_option_parser'
-require 'configuration'
-require 'certificate_generator'
-require 'workers/bulk_sender_worker'
-require 'models/csv_parser'
-require 'models/attendee'
+require 'certificator/configuration/certificate_option_parser'
+require 'certificator/configuration'
+require 'certificator/certificate_generator'
+require 'certificator/workers/bulk_sender_worker'
+require 'certificator/models/csv_parser'
+require 'certificator/models/attendee'
