+++ lib/certificator/configuration/configuration_error.rb
