+++ lib/certificator/converters/inkscape_converter.rb
