+++ lib/certificator/converters/prawn_converter.rb
