+++ lib/certificator/decorators/append_text.rb
