+++ lib/certificator/decorators/identity.rb
