+++ lib/certificator/decorators/multi.rb
