+++ lib/certificator/decorators/prepend_text.rb
