+++ lib/certificator/decorators/replace_text.rb
