+++ lib/certificator/decorators/space_cleaner.rb
