+++ lib/certificator/delivery/aws.rb
