+++ lib/certificator/delivery/dry_run.rb
