+++ lib/certificator/delivery/smtp.rb
