+++ lib/certificator/mailer/certificate_mailer.rb
