+++ lib/certificator/models/attendee.rb
