+++ lib/certificator/models/certificate.rb
