+++ lib/certificator/models/certificate_model.rb
