+++ lib/certificator/models/mailed_certificate.rb
