+++ lib/certificator/workers/bulk_sender_worker.rb
